# backend/openapi.yml
openapi: "3.0.2"
info:
  title: Rest API Docs
  version: "1.0"
servers:
  - url: http://localhost:4444/
  - url: https://ciaro360-server.vercel.app/
components:
  schemas:
    ResponseWithMsg:
      type: object
      properties:
        status:
          type: boolean
        msg:
          type: string
...
    ResponseWithMsgError:
      type: object
      properties:
        status:
          type: boolean
          example: false
        msg:
          type: string
...
    LogsData:
      type: object
      properties:
        _id:
          type: string
        objectType:
          type: string
          enum:
            - policy
            - procedures
            - controls
            - users
            - tags
        objectId:
          type: string
        userId:
          type: string
        action:
          type: string
          enum:
            - add
            - delete
            - update
            - review
            - approve
        description:
          type: string
        timestamp:
          type: integer
        __v:
          type: integer
...
    Outgoing_Tags:
      type: object
      properties:
        listed:
          type: array
          items:
            type: string
          example: ["64410f9eb6f82aada0625a94", "64410f9eb6f82aada0625a94"]
        unlisted:
          type: array
          items:
            type: string
          example: ["Hr", "It"]
    PolicyCreate:
      type: object
      properties:
        projectId:
          type: string
        title:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /logs/policy/{id}:
    get:
      tags:
        - "Log Routes"
      description: "Retrieve logs for a specific policy"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the policy
          schema:
            type: string
      responses:
        "200":
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/LogsData"
        "404":
          description: Logs retrieval failed or policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        "401":
          description: Unauthorized
      security:
        - bearerAuth: []

  # Roles Management Route
...
  /policy/create:
    post:
      tags:
        - "Policy Routes"
      description: "Create a policy"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyCreate"
      responses:
        "200":
          description: Policy created successfully
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "64410f9eb6f82aada0625a94"
                      nameId:
                        type: string
                        example: 101
                  
        "404":
          description: Policy creation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        "401":
          description: Unauthorized
      security:
        - bearerAuth: []
  /policy/{id}/comment:
    post:
      tags:
        - "Policy Routes"
      description: "Add a comment to a policy"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the policy to add a comment to
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                      type: string
                  example:
                    ["64410f9eb6f82aada0625a94", "64410f9eb6f82aada0625a94"]
                comment:
                  type: array
                  items:
                      type: string
                  example:
                    ["64410f9eb6f82aada0625a94", "64410f9eb6f82aada0625a94"]

      responses:
        "200":
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        "404":
          description: Adding comment to policy failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        "401":
          description: Unauthorized
      security:
        - bearerAuth: []
  /policy/{id}/comment/{commentId}:
    delete:
      tags:
        - "Policy Routes"
      summary: "Delete a comment on a policy"
      description: "This endpoint deletes a specific comment on a policy version."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
        - name: commentId
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the comment."
      responses:
        '200':
          description: "Comment deleted successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
  /policy/{id}/reminder:
    put:
      tags:
        - "Policy Routes"
      description: "Set a reminder for a policy"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the policy to set a reminder for
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                next_reminder:
                  type: integer
                  example: 1686720766
      responses:
        "200":
          description: Reminder set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        "404":
          description: Setting reminder for policy failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        "401":
          description: Unauthorized
      security:
        - bearerAuth: []
  /policy/{id}/tags:
    post:
      tags:
        - "Policy Routes"
      description: "Add tags to a policy"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the policy to add tags to
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                tags:
                  $ref: "#/components/schemas/Outgoing_Tags"
      responses:
        "200":
          description: Tags added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        "404":
          description: Adding tags to policy failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

        "401":
          description: Unauthorized
      security:
        - bearerAuth: []

    delete:
      tags:
        - "Policy Routes"
      summary: "Delete tags from a policy"
      description: "This endpoint deletes specific tags from a policy."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
                  description: "List of tags to delete from the policy."
      responses:
        '200':
          description: "Tags deleted successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
  /policy/{id}/procedures:
    post:
      tags:
        - "Policy Routes"
      summary: "Map one policy to many procedures"
      description: "This endpoint maps one policy to multiple procedures."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                procedures:
                  type: array
                  items:
                    type: string
                  description: "Array of procedure IDs."
              required:
                - procedures
      responses:
        '200':
          description: "Mapping successful"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError" 

    delete:
      tags:
        - "Policy Routes"
      summary: "Unmap procedures from policy"
      description: "This endpoint unmaps procedures from a specific policy."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                procedures:
                  type: array
                  items:
                    type: string
                  description: "Array of procedure IDs to unmap."
      responses:
        '200':
          description: "Procedures unmapped from policy successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
  /policy/{id}/projects:
    post:
      tags:
        - "Policy Routes"
      summary: "Map one policy to many projects"
      description: "This endpoint maps one policy to multiple projects."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projects:
                  type: array
                  items:
                    type: string
                  description: "Array of project IDs."
              required:
                - projects
      responses:
        '200':
          description: "Mapping successful"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

    delete:
      tags:
        - "Policy Routes"
      summary: "Unmap projects from policy"
      description: "This endpoint unmaps projects from a specific policy."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projects:
                  type: array
                  items:
                    type: string
                  description: "Array of project IDs to unmap."
      responses:
        '200':
          description: "Projects unmapped from policy successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
  /policy/{id}/controls:
    post:
      tags:
        - "Policy Routes"
      summary: "Map one policy to many controls"
      description: "This endpoint maps one policy to multiple controls."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                controls:
                  type: array
                  items:
                    type: string
                  description: "Array of control IDs."
              required:
                - controls
      responses:
        '200':
          description: "Mapping successful"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

    delete:
      tags:
        - "Policy Routes"
      summary: "Unmap controls from policy"
      description: "This endpoint unmaps controls from a specific policy."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                controls:
                  type: array
                  items:
                    type: string
                  description: "Array of control IDs to unmap."
      responses:
        '200':
          description: "Controls unmapped from policy successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
  /policy:
    get:
      tags:
        - "Policy Routes"
      description: "Retrieve all policies"
      parameters:
        - in: query
          name: page
          required: true
          description: Page number for pagination
          schema:
            type: integer
        - in: query
          name: count
          required: true
          description: Number of policies per page
          schema:
            type: integer
        - in: query
          name: field
          description: Field to sort the policies by
          schema:
            type: string
            default: updated_at
        - in: query
          name: order
          description: Sort order (asc or desc)
          schema:
            type: string
            default: asc
        - in: query
          name: search
          description: search query (search in title and it's id)
          schema:
            type: string
      responses:
        "200":
          description: Policies retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  count:
                    type: number
                    example: 10
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        ID:
                          type: number
                        
                        title:
                          type: string
                        
                        tags:
                          type: object

                        assignees:
                          type: array
                          items:
                            type: string
                        
                        created_by:
                          type: string
                        
                        updated_at:
                          type: string
                          format: date
                          
                        status:
                          type: string
                        
        "404":
          description: Policy retrieval failed or policies not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        "401":
          description: Unauthorized
      security:
        - bearerAuth: []
  /policy/overview:
    get:
      tags:
        - "Policy Routes"
      description: "Get policies overview"
      responses:
        "200":
          description: Overview retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      "Total Count":
                        type: integer
                      Active:
                        type: integer
                      Inactive:
                        type: integer
                      "In Draft":
                        type: integer
                      "In Review":
                        type: integer
                        
        "404":
          description: Policy retrieval failed or policies not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        "401":
          description: Unauthorized
      security:
        - bearerAuth: []
  /policy/fromTemplate:
    post:
      tags:
        - "Policy Routes"
      summary: "Create a policy from a template"
      description: "This endpoint creates a policy from given templates."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId:
                  type: string
                  description: "The ID of the project."
                templateIds:
                  type: array
                  items:
                    type: string
                  description: "Array of template IDs."
                assignees:
                  type: object
                  properties:
                    author:
                      type: array
                      items:
                        type: string
                    reviewer:
                      type: array
                      items:
                        type: string
                    approver:
                      type: array
                      items:
                        type: string
                  description: "list of assignee IDs."
              required:
                - projectId
                - templateIds
                - assignees
      responses:
        '200':
          description: "Policy created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
        '404':
          description: Not found
          content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ResponseWithMsgError"
  /policy/create/file:
    post:
      tags:
        - "Policy Routes"
      summary: "Create a policy from an uploaded file"
      description: "This endpoint creates a policy from an uploaded file."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: "The file to upload and convert."
      responses:
        '200':
          description: "File processed and policy created successfully"
          content:
            text/markdown:
              schema:
                type: string
        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  msg:
                    type: string
...